[task] move aws specific definitions under tasks/aws
