+++ tasks/aws/eks.py
@@ -10,6 +10,8 @@ from invoke.tasks import task
 from pydantic import ValidationError
 
 from tasks import config, doc, tool
+from tasks.aws import doc as aws_doc
+from tasks.aws.common import get_aws_wrapper
 from tasks.aws.deploy import deploy
 from tasks.destroy import destroy
 
@@ -27,7 +29,7 @@ scenario_name = "aws/eks"
         "linux_arm_node_group": doc.linux_arm_node_group,
         "bottlerocket_node_group": doc.bottlerocket_node_group,
         "windows_node_group": doc.windows_node_group,
-        "instance_type": doc.instance_type,
+        "instance_type": aws_doc.instance_type,
     }
 )
 def create_eks(
@@ -93,7 +95,7 @@ def _show_connection_message(ctx: Context, full_stack_name: str, config_path: Op
     except ValidationError as e:
         raise Exit(f"Error in config {config.get_full_profile_path(config_path)}:{e}")
 
-    command = f"KUBECONFIG={kubeconfig} {tool.get_aws_wrapper(local_config.get_aws().get_account())} kubectl get nodes"
+    command = f"KUBECONFIG={kubeconfig} {get_aws_wrapper(local_config.get_aws().get_account())} kubectl get nodes"
 
     print(f"\nYou can run the following command to connect to the EKS cluster\n\n{command}\n")
