+++ tasks/azure/doc.py
@@ -1,0 +1,1 @@
+instance_type: str = "The instance type to use (default is Standard_B4ms for azure)"
